fix requirement
