+++ requirements.txt
@@ -2,5 +2,4 @@ numpy
 tqdm
 openai
 elasticsearch
-sentence_transformers
-import torch
+torch
